+++ index.html
@@ -13,7 +13,7 @@
       <h3 class="title is-3">Vertices</h3>
       <form id="createVertices" autocomplete="off">
         <div class="input-field">
-          <input id="vertices" name="vertices" type="number" class="input is-primary" value="6">
+          <input id="vertices" name="vertex" type="text" class="input is-primary" value="" placeholder="Vertex">
         </div>
         <div class="">
           <button type="submit" class="button is-primary is-outlined">Create</>
@@ -30,7 +30,7 @@
     <div class="column">
       <div class="columns">
         <div class="column">
-        </br>
+          </br>
           <button id="show-graph" class="button is-primary is-outlined">SHOW ADJACENCY MATRIX</button>
           <div class="columns">
             <div id="canvas" class="column">
