+++ src/app/graph.js
@@ -1,0 +1,11 @@
+class Graph {
+  constructor(adjacencyMatrix, vertices) {
+    this.adjacencyMatrix = adjacencyMatrix;
+    this.vertices = vertices;
+    this.edges = [];
+  }
+  jsonToAdjacencyMatrix(json){
+
+  }
+}
+export default Graph;
