+++ src/index.js
@@ -1,12 +1,14 @@
 import JsonGraph from './app/jsonGraph';
-import WindowEvents from './view/windowEvents';
-import ButtonEvents from './view/buttonEvents';
+import initViewEvents from './view';
 
 let graph = new JsonGraph();
 window.graph = graph;
+if (localStorage.getItem('graph') === null){
+  localStorage.setItem('vertices', ['a']);
+  localStorage.setItem('graph', ['a']);
+}
 
-WindowEvents();
-ButtonEvents();
+initViewEvents();
