+++ src/view/index.js
@@ -1,0 +1,7 @@
+import initEvents from './events';
+
+const initViewEvents = () => {
+  initEvents();
+};
+
+export default initViewEvents;
